pyproject.toml
[tool.poetry]
name = "llm-backend"
version = "0.1.0"
description = ""
authors = ["rishad.m <rishad.m@mitz.ai>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = ">=3.10"
fastapi = ">=0.115.12,<0.116.0"
uvicorn = {extras = ["standard"], version = ">=0.34.1,<0.35.0"}
python-dotenv = ">=1.1.0,<2.0.0"
elasticsearch = ">=9.0.0,<10.0.0"
requests = "^2.31.0"
pdf2image = "^1.16.3"
pillow = "^9.5.0"
openai = "^1.75.0"
python-multipart="^0.0.9"

[tool.poetry.dev-dependencies]
black = ">=25.1.0,<26.0.0"
pre-commit = "^3.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
